v4l-utils: create separate keymaps instead of patching rc6_mce
All locally added remote keycodes are now in separate files,
one per remote, instead of adding them to rc6_mce.

A new keymap, libreelec_multi, with all keycodes of the previous
rc6_mce map is added in install and used by default when the
remote is configured to use the rc-rc6-mce table.

The default behaviour is unchanged, LE can still be controlled with
MCE, xbox, ... remotes out of the box but it's now possible to configure
the ir receiver to react only on the codes sent by a specific remote.

This solves the issue where a xbox remote always controlled LE even
when a user only wanted to control it with a MCE remote.

To do this simply create a /storage/.config/rc_maps.cfg file with
the following content:

*	rc-rc6-mce	rc6_mce

Signed-off-by: Matthias Reichl <hias@horus.com>