Fix for Issue 113 [Compiling Lakka for Cubox fails @ Building Retroarch]
i dug a little deeper and I found a solution to the issue.
packages/libretro/retroarch/package.mk was missing the correct $OPENGLES param for the GPU on a Cubox-i 4 Pro.
"imx-gpu-viv" is the correct string for this board.
Maybe there should be a check here that informs someone that $OPENGLES is set to an unknown parameter?